+++ src/main/java/com/github/davidmc24/gradle/plugin/avro/AvroBasePlugin.java
@@ -24,6 +24,7 @@ public class AvroBasePlugin implements Plugin<Project> {
         configureExtension(project);
     }
 
+    @SuppressWarnings("deprecation")
     private static void configureExtension(final Project project) {
         final AvroExtension avroExtension =
             GradleCompatibility.createExtensionWithObjectFactory(project, Constants.AVRO_EXTENSION_NAME, DefaultAvroExtension.class);
@@ -38,8 +39,11 @@ public class AvroBasePlugin implements Plugin<Project> {
             task.isGettersReturnOptional().convention(avroExtension.isGettersReturnOptional());
             task.isOptionalGettersForNullableFieldsOnly().convention(avroExtension.isOptionalGettersForNullableFieldsOnly());
             task.isEnableDecimalLogicalType().convention(avroExtension.isEnableDecimalLogicalType());
+            task.getConversionsAndTypeFactoriesClasspath().from(avroExtension.getConversionsAndTypeFactoriesClasspath());
             task.getLogicalTypeFactories().convention(avroExtension.getLogicalTypeFactories());
+            task.getLogicalTypeFactoryClassNames().convention(avroExtension.getLogicalTypeFactoryClassNames());
             task.getCustomConversions().convention(avroExtension.getCustomConversions());
+            task.getCustomConversionClassNames().convention(avroExtension.getCustomConversionClassNames());
         });
     }
 }